requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argilla==1.3.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.25
click==8.1.3
clickhouse-connect==0.5.25
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.14
docarray==0.32.1
duckdb==0.8.0
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.95.2
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.13.2
jedi==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.180
Levenshtein==0.21.0
llama-index==0.6.11
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
onnxruntime==1.15.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.9.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
parso==0.8.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
posthog==3.0.1
prompt-toolkit==3.0.38
protobuf==4.23.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.8
Pygments==2.15.1
pypandoc==1.11
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-Levenshtein==0.21.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.0.0
regex==2023.5.5
requests==2.29.0
rfc3986==1.5.0
rich==13.4.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
traitlets==5.9.0
typer==0.9.0
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
unstructured==0.7.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
